requirements.txt
Pillow~=9.3.0
requests~=2.28.1
transformers~=4.30.2
torch~=2.0.1+cu118
torchvision~=0.15.2+cu118
numpy~=1.24.1
matplotlib~=3.7.2